requirements.txt
streamlit
Pillow
fpdf